requirements.txt
langchain==0.2.16
langchain-openai==0.1.23
langchain-community==0.2.16
langchain-core==0.2.38
chromadb==0.4.22
tiktoken==0.7.0
python-dotenv==1.0.0
pydantic==2.8.2
